Adds and Defines FeaStiffenerArray
This creates the class FeaStiffenerArray, defined as an
array of FeaSlices perpendicular to the spine of a geometry
with only cap beam elements included. In Update, the number
of stiffeners is first calculated. This number is iterated over in
CreateFeaStiffenerArray to call ComputeSliceSurf while updating
the center location and saving the resulting surfaces. The function
FeaPartIsArray is addded to identify if an FeaPart is either an
FeaRibArray or FeaStiffenerArray.