Updates and Improves FeaSlices
First, additional orientation options are added for FeaSlices to be
defined from the body coordinate frame of the parent geom or the
absolute x,y,z axes. The spine of the parent geometry is now used to
define a slice that runs perpendicular to the spine. Visualization of
the spine is added for debugging purposes. The function UpdateParmLimits
is added, allowing its functionality to be removed from
ComputeSliceSurf. Last, the GUI is updated for these changes.