Creates SSLineArray Class in SubSurface
This adds a new type of SubSurface, SSLineArray. This class
will primarily be used be FeaStructure to create an array of
beam elements along constant U or W lines. First, the number
of Lines is calculated. Then, m_LVec is developed similar to
SSLine::Update to create an array. SSLineArray is also added
to APIDefines and ScriptMgr