BUGFIX: Solves Issue with Update Loop for Arrays
An issue was identified with setting Parm values for an FeaArray
from within their Update or ComputeSurf functions. Setting the
Parm value was causing FeaPart::ParmChanged to be called and
all screens to be updated. Updating the screens would then cause
the FeaStructure to update, the FeaArray to update, and the cycle
to repeat itself. This commit solves the bug removing calls to set
the Parm values and updating the associated FeaArray functions
accordingly.