BUGFIX: Fixes Connectivity Error in FEA Meshes
In FeaMeshMgr::BuildFeaMesh(). The FeaTris were defined for each surface
in m_SurfVec individually, which caused the intersection nodes to be
defined multiple times. When performing an FEA analysis, the FeaParts
were not dispaced together. Similar to some code in
MeshGeom::BuildIndexedMesh, all of the tris and nodes are collected and
NanoFlann is used to determine coincident points. The indexing of
the Tris is updated prior to building the FeaMesh.