Improves FeaSpar::ComputePlanarSurf() with End Cap Considerations
Instead of using surface U and W values to beuild the FeaSpar
surface, the location along the aferage chord line is used. This
allows for the parameterization by chord length or percent along
chord to correctly center the spar. End Caps are taken into
consideration to correctly determine the start and end points of
the FeaSpar.