BUGFIX: Adds Connectivity for FeaBeam Elements
FeaBeam elements are created in BuildFeaMesh() using the m_TessVec
of the each index of m_ISegChainList. However, the points in m_TessVec
were often slightly different than the previously defined triangle nodes.
The tolerance in BuildIndMap and FindPntInd was causing similar nodes
to be indexed differently. This adds a check to compare the distance
between FeaBeam endpoints and existing triangle nodes and use the
triangle node value if similar.