Changes FeaMaterial and FeaProperty Indexes from Ints to IntParms
This changes m_FeaMaterialIndex and m_FeaPropertyIndex for
for FeaParts and SubSurfaces from ints to IntParms. This is mainly
done to ease setting of the FeaMaterial index and FeaProperty index
though the API and ScriptMgr. Also, this eliminates the need for
helper functions to get and set the indexes. As parms, the no longer
need to be included in Encode and Decode XML functions.