Moves Enums from FeaStructure to APIDefines and Improves Names
This takes the FeaSlice orientation enums and the included elements
enums in FeaStructure and adds them to APIDefines for use through
the API. The names of the included elements are modified to improve
clarity. All cases where the enums are used are update with correct
names and the addition of the vsp:: class prefix.