Adds API Functions for FeaStructures to VSP_Geom_API
This adds a series of functions for using the API to perform the same
functionality available with StructScreen and FeaPartEditScreen.
The functions included enable adding FeaStructures, FeaParts,
FeaSubSurfaces,FeaMaterials, and FeaProperties. The StructSettings
and FeaGridDensity can be modified and the export file names can
be set. Last, an FEA Mesh is generated and exported with
ComputeFeaMesh.