Adds FEA Mesh API Test
TestFEAMesh() is added to APITestSuite to test and demonstrate
many of the features availble through the StructScreen and
FeapArtEditScreen. A pod is added and an FeaStructure is created
with it. An FeaMaterial and FeaProperty are created and a series
of FeaParts and FeaSubSurfaces are added to the FeaStructure.
The vehicle is saved, the Calculix export file is identified, and the
FEA Mesh is generated.