Adds All FeaSlice Options to FeaSliceArrays
This makes available many more options for FeaSliceArrays, such
as specification of orientation plane and rotation.
m_IncludedElements is no longer set to vsp::FEA_BEAM only.
CalcNumSlices is updated to consider the orientation of the FeaSlices.
GUI elements are added to FeaPartEditScreen to provide user control
of the new options.