BUGFIX: SubSurface Edge Split on Border Curve Cap Identification
FeaSubSurfaces with an edge on a border curve that is split by an FeaPart
were not being identified correctly, causing lost beam elements. First,
m_CompID is added as an input to GetSimpSubSurfs, where no
SubSurfaces are returned if m_CompID < 0, indicating an FeaPart surface.
SubSurfaces on FeaParts are not supported, and were being misidentified.
CheckSubSurfBorderIntersect is updated with additional checks and
corrections for missing parentheses.