Separates FeaPart Locations Parms to Absolute or Relative
First, the LENGTH and FRACTION enums are repaced with the use of
vsp::ABS and vsp::REL. Similar to how Geom XForm transformation Parms
are separated by absolute and relative parameterization, the same
approach is used with FeaPart location Parms. The inactive Parm is
updated and functions are modified to only use the relative location
Parm value.