Removed Display and Tagging of FeaNodes for SubSurface Holes
First, SubSurface nodes are not tagged if the included elements are
BEAM, since the tris are deleted in RemoveSubSurfFeaTris. In
LoadDrawObj, FeaNode tag is used for identification, instead of index.
This prevents nodes from deleted tris to be visible and allows nodes
with more than one tag to be visible.