OpenVSP Version 3.13.2
Just in time for the workshop an update with some features and lots of bug
fixes.

This adds a built-in body of revolution -- in case the Duct custom
components were not enough for your needs.  Also includes Spin control for
fuselage and stack geoms.

Lots of updates to VSPAERO -- ground effect and some viewer updates, but
mostly a whole host of bug fixes for issues that people have been running
into.

Plenty of good stuff here.  Everyone should update.

Features:
 - Body of revolution component
 - Spin control for fuselage and stack
 - VSPAERO Update to 4.1.1
  . Added ground effect model
  . Reflect symmetric half geometry in viewer
  . Show XYZ planes in viewer

Bug Fixes:
 - Clean up logic in STEP/IGES export - may help some importers
 - VSPAERO GUI now creates half-mesh for panel method with symmetry
 - VSPAERO Update to 4.1.1
  . Fix drag integration inconsistency for alpha vs. rotation
  . Fix memory crash w/ periodic wakes (ducts)
  . Fix pressure anomaly w/ periodic wakes (ducts)
  . Fix problem with symmetry calculations
  . Fix high AR Cp near centerline
  . Fix high AR Cp out span
  . Fix slicer on Windows
  . Set minimum Mach to 0.001 -- fixes Cp min/max range in viewer