Force update of children of hinges to correct initial position.
Normally, Geoms are located at the origin when they are added.  So,
the attachment matrix doesn't do anything.

The attach matrix is built by knowing the ParentID, but that isn't set
until after the last Update() in the normal process.

For the case of a hinge parent, we force attachment.  So, we need another
Update() after ParentID() is set -- in Vehicle::AddGeom( Geom* ) called
at the top of Vehicle::AddGeom( const GeomType & ).