Improves Formatting of FEA Output Files, Particularly for NASTRAN
The function GetFeaFormat is added to return a string format for an
input double value. This function is used to improve the formatting
of the NASTRAN output file, in which a maximum of 8 characters are
allowed for each data entry. Instead of using CBEAM elements and
PBEAM properties for NASTRAN, CBAR and PBAR are now used due
to their preferred simplicity. This commit also includes misc. cleanup
of comments.