Adds Beam Cross Section Support to SimpleFeaProperty and Export Files
The cross section type and dimensions are copied from FeaProperty to
SimpleFeaProperty. In WriteCalculix and WriteNASTRAN, the section
properties written to file now differ if the cross section type is not
General. The NASTRAN dimension specification for each cross section is
used as the standard, and Abaqus/Calculix cross section dimensions are
interpreted from these standards.