Adds Beam Cross Section Support to StructScreen
This adds GUI elements unique to each cross section type and updates
the beam cross section parms appropriately. The FeaProperty Browser
is also updated to list the cross section of each FeaProperty. NOTE:
"Show XSec" button is not supported at this time, but will be used to
present diagrams of the XSecs.