Adds Option to Attach Dome Center to Parent Geom Spine
A BoolParm flag is added to FeaDome to indicate whether or not the
initial dome center should lie on the spine of the parent geom. Another
parm is added to specify the distance along the spine the dome center
is located. This feature makes FeaDomes more parametric. GUI elements
are added to FeaPartEditScreen.