Adds Function for Merging Coplanar FeaParts
This function checks all FeaPart surfaces prior to being loaded in
AddStructureParts to determine if they are coplanar. If they are
coplanar and their bounding box intersect, the coplanar surfaces are
deleted and a new surface is created. When this occurs, FeaPart data
is only kept for the earlier FeaPart index in for the FeaStructure.