Improves CheckFixPointIntersects
First, the tolerance is lowered slightly if the fixed point is on an
FEA part surface, since it is more difficult to identify if the fixed
point is on an intersection curve. Next, initial guesses are added to
ClosestUW to improve speed and accuracy. Last, a check is added to
identify if m_SplitVec is increased after AddBorderSplit is called.