Splits Wing Section Constraint for Spars into Start and End Section
Instead of only having the option to constrain a spar to a single wing
section, the start and end wing section can now be specified. The
FeaPartEditScreen is updated to support the addition. Also, tracking of
Umin and UMax for the spar wing section is added as a member variable to
reduce code duplication.