Moves Rib Functions From FeaPart to FeaRib and Updates
Since these functions are only used by FeaRib and FeaRibArray, it is
more reasonable for them to be located in the FeaRib class. They were
added to FeaPart originally due to a parm update bug that has since
been corrected. FeaRibArray is updated to add a temporary rib, set
parms, and create a surface from it instead of using the FeaPart
functions. Member variables are added to FeaRib to reduce code
duplication.