Adds Option to Specify a Start and End Wing Section to Constrain a Rib
A parm is added to FeaRib that is used to indicate if sectional
control shall be used. If true, the indicated start and end wing
sections are used to determine the beginning and ending U values when
building the trapezoidal wing box. The FeaPartEditScreen is updated
with GUI elements to support the added capability.