Adds Reset Display Feature
First, a simple function is added to FeaMeshMgr to indicate if
FeaMesh data is available. ShowAllParts is added to StructureMgr, which
shows only the selected structure's parent geom and all of the FeaParts
and FeaSubSurfaces. In StructScreen, a button is added to call
ShowAllParts and set all mesh display flags to false.