BUGFIX: res->GetNumData Replaced by GetDoubleData().size()
GetNumData in NameValCollection was being used in VSPAEROPlotScreen
to determine the number of x and y sweep data values (same as number of
wake iterations), but GetNumData returns the size of the NameValData
vectror assocciated with the data name. This was always returning 1,
causing the first wake iteration result to be plotted. Using
GetDoubleData().size() fixes this problem.