Updates FeaSlice::ComputeSliceSurf to Resize for XYZ Center Offsets
By default, the center location of an FeaSlice is always aligned with
at least two axes of the parent geom central axes. To allow FeaSlice to
be used as a rib or spar, center offsets need to be considered. This is
because the center of a rib or spar is based off the wing geometry, not
the bounding box.