Updates FeaRib Definition and Adds Bounding Box Trim Option
First, Ribs are updated to be defined initially from the pre-transformed
parent geom. ComputeRibSurf is reorganized and tracking of the wing
normal vector is added to reduce code duplication. Last, a BoolParm
flag is added to provide the option of trimming a rib to the section(s)
bounding box. If this flag is true, a temporary FeaSlice is used to
calculate the rib surface after the rib center and section bounding box
are set for the slice.