Adds Feature Line Suppression to Fixed Point Surface Identification
The call to IdentifySplitSurfIndex is moved from Update in FeaFixPoint
to AddStructurePart in FeaMeshMgr, since it is only needed just before
mesh generation. Tracking of the half mesh flag in update functions is
removed and replaced as an input to IdentifySplitSurfIndex. The vectors
of suppressed U and W feature lines are added as inputs as well,
allowing for correct identification of Fixed Point parent surface(s)
after feature line suppression.