Adds Function to Merge Overlapping FeaPart & SubSurface Intersections
This function is used to check if a subsurface edge will be exactly
aligned with an FeaPart intersection curve. This is done by checking the
distance between the subsurface edge endpoints and the closest points on
the FeaPart surface. If the subsurface edge lies on the FeaPart surface,
its intersection chain is found and removed, since the FeaPart will
create create a similar intersection curve with the skin. This is done
to avoid overlapping and disconnected FeaElements.