Fixes Absolute Slice Location Update by Reorganizing Center Calculation
Calculation of the slice center relies on the setting of the relative
location parm. This was previously causing an update lag for absolute
location changes. The order is reorganized to determine the bounding box
distance perpendicular to the slice orientation, set the relative or
absolute parm values, and then calculate the slice center.