Moves Setting of XYZ Rotations to FeaSliceArray Class and Fixes Update
Similar to changes made earlier for FeaSlice rotation settings, they
are moved from FeaPartEditScreen to FeaSliceArray. In addition, the call
to UpdateParmLimits for each slice in the array is moved to after the
setting of the relative location parm. This order is necessary to
identify the 3D center of each slice in the array.