Reduce initial memory (and increment) for VBO allocations.
These parameters dictated that even a trivial graphics object required
substantial memory.  A smarter system could be devised where large
buffers are extended in larger steps, but simply targeting the size
closer to the requirement is an improvement.  Though theroretically
slower, it will be much more efficient in memory use.