Adds Support for CpSlicer Result Plotting
This adds a new tab to the VSPAEROPlotScreen for displaying
CpSlicer results. Results from the CpSlicer are identified by a
cut (CpSlice pointer info) and flow condition case. GUI elements
and update functions are added to identify the cut and case
names and selected result IDs, from which the data to be plotted
is identified.