Adds XYZ Position Data Plotting Specification
The existing CpSliceCutBrowser is divided in half so that a choice can
be added for each cut to identify the position data to plot. A map is
added to track all result IDs corresponding to each cut index.
Another vector is added to track the position type for each cut index.
The x axis label is updated to identify the position type being plotted.