Adds Visualization of CpSlice Cut Planes
First, a BoolParm is added to CpSlice that indicates whether or not to
draw the cutting plane. CreateSurf creates the cutting plane surface by
using MakePlaneSurf. It is called in LoadDrawObj to define the lines
used to draw the plane. Two DrawObjs are used, one for the plan surface
another for the outline, to avoid transparancy ordering issues. In the
VSPAEROScreen, a toggle button is added to update the CpSlice drawing
flag.