Identifies Cp or dCp from Analysis Type
The analysis type for the CpSlicer is saved in UpdateFilenames,
to distinguish it from the current value of m_AnalysisMethod,
which may not be the same as the value in the most recent
VSPAERO setup file. The analysis type is used to identify if dCp
or Cp values are saved as results and added to the plot.