Makes X-Axis Position Specific for Slice Type, Instead of Each Slice
This removes the vector of choices that corresponded to the x-axis
position type for each slice. It is replaced by three choices specifying
the position type for X, Y, and Z cut types. The choices are activated
or deactivated based on which cut types are available. The choices are
also used to identify the x-axis data to plot in RedrawCpSlicePlot.