Adds Support for Fixed Points on Merged FeaParts at the Same Location
When a fixed point is on an FeaPart that is placed in the same location
and oriented the same as another FeaPart, it will be moved if its parent
is lost in the merge. Note: FixedPoints on merged FeaParts due to
symmetry is not supported, since the UW coordinates of the new oversized
merged surface will not map correctly.