BUGFIX: Eliminates Half Mesh Collapsed Beam Elements
First, an additional consideration for FEA parts is added to
CleanMergeSurfs to increase the tolerance in SurfCore::LessThanY.
This compensates for the oversizing done for FEA parts. The call
to CleanMergeSurfs is also moved to after FEA Part surfaces are
loaded. In addition, a check for collapsed beam elements is added
to BuildFeaMesh to prevent beam elements with the same start
and end node from being created.