Fixes Updating and Activation of Default FeaMaterials
An Update function is added to FeaMaterial, and is called in
ParmChanged. In Update, the parm values for the default materials are
set, instead of being set in StructureMgrSingleton::InitFeaMaterials.
This prevents the parm values for the default materials from being
changed even when their associated GUI devices are deactivated.