Adds Selig Airfoil Export Example Script
This script is used to calculate the coordinate points of a VKT airfoil
and write the data to a Selig airfoil file. The output of this script is
loaded in XFoil to generate the Cp distribution. The Cp distribution is
saved to a file that is called in the Master V&V script to compare
CpSlicer results to XFoil for the VKT airfoil.