Adds VSPAERO Preconditioner Choice
Enums for the 3 Preconditioner types are added to APIDefines and
registered in ScriptMgr. m_JacobiPrecondition BoolParm is changed to an
IntParm and renamed to m_Precondition. The default is set to matrix
precondition. If the preconditioner is set to Jacobi, the "-jacobi"
argument is input to VSPAERO. If it is set to SSOR, the "-ssor" input
is sent.