Adds API Function to Calculate Hershey Bar Wing Lift Distribution
This function calculates the lift distribution across a Hershey Bar wing
with unit chord length using Glauert's method. The inputs are the number
of points across the half-span in which to evaluate, angle of attack,
free stream velocity, and total wing span. An aditional bool input is
available if the lift distribution across the entire span is desired.
The returned vector<vec3d> contains each Y evaluation point and
corresponding Cl value.