Adds VKT Airfoil Data and XFoil Cp Solution
The first file, VKT_e0.1_k0.1_t10.dat is a Selig airfoil file generated
by calling vsp::WriteSeligAirfoilFile with the VKT airfoil coordinate
points determined by vsp::GetVKTAirfoilPnts. This airfoil file is loaded
in XFoil, the inviscid solution is computed, and the result is saved to
XFoil_VKT_CpDist.txt. This file will be read in the Master V&V script to
provide a comparison between VSPAERO and XFoil for the VKT solution.