Adds and Registers GetSubSurf Overload Function
This function returns a vector of subsurface IDs for the given geom ID
and subsurface name. In Script Mgr, a wrapper function is added to
convert the returned vector<string> to a CScriptArray*. In addition,
the registration of the existing GetSubSurf function is is updated to
use "asFunctionPR" to support the function overloading.