Implement U/W Line Sources
This allows a hybrid XYZ and UW neighborhood search for U/W constant line
sources.  This improves the resulting quality without dramatically
increasing the number of line segments used to represent them.

This is implemented with a layering violation to allow the source
to evaluate in u/w space as well as x,y,z.