Clean up UpdateDegenDrawObj
First, change size indicator to the actual vector that is referenced, not
just another vector that should have the same size.

Second, remove reference to m_SurfVec to check the normal vector direction.
Could possibly simply use approx_norm in place of camber_draw_norm, but
to keep things as the same as possible, just use it to compare for
parallel directions.