OpenVSP 3.14.0
'Tis the season for OpenVSP Holiday Pi.  Partake heartily.

It has been far too long since the last release.  This one started out
small, but then had to wait for some things to finish.  While waiting,
it grew.  Then it had to wait on something else -- and it grew more.  Now
that it is finally out the door, it is no longer a small update.

Lots of great features and bug fixes for both VSPAERO and OpenVSP.  Too
many things to highlight just one or two.  Download the latest, check out
the release notes, and enjoy.

Features:
 - VSPAERO Update to 4.3
  . Matrix based preconditioner is now default
  . Vortex lift on by default
  . Leading edge suction a separate option
  . Option to disable 2nd order Mach correction
  . Multiple Re support for CDo (not supported yet by OpenVSP GUI)
  . Some small changes to calculating the surface pressures on the body.
 - OpenVSP updates to support VSPAERO 4.3
 - CPSlicer GUI
 - VSPAERO Automated Test Script
 - SubSurface API Extensions
 - Improvements to DegenGeom representation of control surfaces
 - Wireframe import of *.hrm files for CompGeom/DegenGeom based analysis
 - Measure, point probe in GUI
 - Surface query API routines
 - U,W Line Sources for CFDMesh
 - Some outputs now handled as Parms and can be used in linking.
 - Defer Update() in CreateGeom

Bug Fixes:
 - First (instead of last) wake iteration plotted in VSPAEROPlotScreen
 - Fix VBO buffer size and increment to prevent crashing on complex models
 - Scale rounded rectangle corner radius.
 - Scale root section of multi-section wings.
 - Remove dangling panel method mesh every time VSPAERO is initiated.
 - Fix counter-intuitive behavior pasting components onto hinges.
 - Fix hinge GUI name update.
 - Misc. other bugs
 - VSPAERO 4.3 Bug fixes
  . Fixed a bug with forces/moments for propulsive cases - vlm and panel
  . Various other bug fixes..