Don't convert surface to cubic before creating patches.
Patches were only used for intersection curves and interior test.  The
old code for these was hard-coded to bi-cubic Beziers.  Since this was
converted to Code-Eli surfaces, this can now be generalized.

For this to be robust, it requires either more refinement levels in
the intersection calculation or an adaptive algorithm.