OpenVSP 3.14.2
Mostly a bug fix release, but a few treats thrown in for good measure.

This release fixes a bunch of issues that cropped up in 3.14.  So far, no
problems have been reported that relate to the structures support in .1,
so only one release will be made here.

In addition to the fixes, two 'features' are included.

First, we're including pre-built Python bindings!  Nothing here that
you couldn't do yourself before, but now that we're doing the heavy
lifting, a bunch more people should be able to Python their OpenVSP
to their hearts content.  These bindings will only work with the
Python they were built against.  These were built against Anaconda3.6.
(The 32-bit version on Windows.)

Second, there is an update to libIGES that allows naming of the exported
surfaces.  We use Entity 406 Type 15 names -- this allows names to be
of arbitrary length.  You might want simple names derived from the
component names -- or you might want something guaranteed unique that
identifies all split surfaces with their parent - and symmetrical copies
with their master.  The IGES export GUI now has a bunch of optinos to allow
control of the name fields and delimeter.

The MacOS release has been built with a new compiler on a newer machine,
please report any improvements or problems.

The fixes are important too, a couple of nasty ones in there.  You'll want
to update.

Features:
 - Python bindings now included - Anaconda3.6 based.
 - IGES surface names

Bug Fixes:
 - Parasite drag SWet showing as zero (thanks Corrado)
 - Fix crash when pasting components to top level (thanks Jasper)
 - Fix VSPAERO GUI layout problems (thanks Marco)
 - Fix problem reading line subsurfaces from file (thanks Nat)
 - Scattered other fixes