Adds Unsteady Tab to VSPAEROPlotScreen
This tab is added to display the time history of unsteady analysis
results, as opposed to the wake iteration history on the Convergence
tab. The convergence tab is deactivated and results are not plotted for
unsteady analysis. Parms are added to VSPAEROMgr to adjust the axis
limits on the Unsteady Tab. Similar to the Convergence tab, functions
are added to update the flow condition browser, Y data browser, axis
limits, and plot.