OpenVSP Version 3.15.0
Happy New Year!  And Happy New OpenVSP!  One significant new feature and
a bunch of fixes and cleanups make this release worthy of a .0 milestone.
I hope to see you at SciTech or AHS this year.

The new feature is support for stand-alone calculation of intersection
curves (without running CFDMesh).  The curves can be visualized and
exported to GridTool and Plot3D formats.  These changes also afforded an
audit of the surface intersection code -- resulting in many improvements.

These improvements should significantly improve the quality of the curves,
which will hopefully improve the robustness and quality of the CFDMesh and
FEAMesh processes.

Unfortunately, the improved algorithms are potentially noticeably slower.
If you notice CFDMesh pausing for a longer period during 'Intersect', this
is why.

There are a bunch of changes to the organization of the Structures data.
This should make all of the Parms available through the API or as design
variables.  However, this will also likely mean that structures currently
saved to a *.vsp3 file will not read into versions going forward.

We try to avoid this sort of thing, but this one slipped by -- very sorry
for the inconvenience.  However, I am certain that the best time to fix
this issue is now.

The VSPAERO GUI has been extended to support unsteady stability runs.

There are a handful of bug fixes.  Most significantly, a bunch of fixes
related to saved views, the openness of the adjust view GUI, and the
preset views from the View menu.

Overall, lots of good stuff.  Download right away.  Enjoy in moderation.

Features:
 - Stand-alone curve intersection generation, visualization, and export
 - VSPAERO GUI integration of PQR unsteady stability stuff

Bug Fixes:
 - Structures data organization to make Parms available to API
 - View adjust bugs