BUGFIX: Removes Buggy stringstream and Replaces with MessageMgr::Send
m_OutStream is removed from SurfaceIntersectionMgr.h and addOutputText
is updated to divert input messages to the MessageMgr. m_MessageName is
used to identify which screen to send the message to
("SurfIntersectMessage", "CFDMessage", or "FEAMessage"). This allows the
monitor functions in CFDMeshScreen, StructScreen, and
SurfaceIntersectionScreen to be simplified significantly. This fixes
issues of messages not showing up in the text output of each screen.