Adds Option to Enable\Disable Exporting of ParasiteDrag Sub-Components
A BoolParm is added to ParasiteDragMgr to indicate whether or not to
include sub-components in the *.csv export file. Its default value is
false. ParasiteDragMgrSingleton::ExportToCSV() is modified to check the
value of the BoolParm. If false, results are only created for the
primary components. A toggle button is added to the ParasiteDragScreen
to set the value of the BoolParm.