Adds WriteBezierAirfoil in Geom
This function is used to write out a single Bezier airfoil file defined
at the input u location on the foil surface. The vector of
BezierSegments is obtained for the constant u curve on the untwisted
unit length foil surface. For each BezierSegment, the order, t0, tend,
and all control points are written to file.