Adds Bezier Control Pnts to Coordinate Pnts Matlab Functon
The Matlab function BezierCtrlToCoordPnts evaluates the Bezier curve
defined in the input file at the input t values (0 <= t <= 1). The
function loads the Bezier airfoil file and identifies a matrix of all Bezier
segment data. For each t value, the corresponding Bezier segment is
identified. The local t value is calculated and used to evaluate the Bezier
segment based on the segment's order. At this time, only segment
orders 1-3 are supported.