Adds Full File Path Consideration for Individual Airfoil Files
Only the name and file directory for the airfoil metadata file is
able to be selected, and all individual airfoil files were
previously being written to the VSP excecutable directory. This
change strips the file directory from the metadata file name
and saves it to a string in Vehicle. In WriteAirfoilFiles, this
file path is added as a prefix to each airfoil file name to ensure
they are placed in the same directory as the metadata file.