Adds SET Definitions and NASTRAN Key File Writing in WriteNASTRAN
A NASTRAN file is divided into multiple sections in which the bulk data
section was the only section previously written to the _NASTRAN.dat
file. NASTRAN SETs are defined in the Case Control section, which occurs
before the bulk data. To allow each element and grid ID to be written in
the Case Control section, a temporary file is used to write all bulk
data while the SET definitions are added at the top of the NASTRAN file.
Once all of the SETs are defined, the temporary file bulk data is read
and copied to the NASTRAN output file.