Adds GetHersheyLLTData Function, LLTData Struct, and CDi Calculation
First, a struct, LLT_Data, is defined in VSP_Geom_API.cpp only so it can
not be used outside of the *.cpp file. It contains vectors of y span
location, circulation, Cl, etc. A function, GetHersheyLLTData, is then
defined in VSP_Geom_API.cpp to calculate all members of LLT_Data given
an input number of points, angle of attack (rad), Vinf, and span. This
allows the existing GetHersheyBarLiftDist to be condensed significantly.