BUGFIX: Fixes Parasite Drag Geom Set Visibility
In SetupFullCalculation, the functions that create the DegenGeom and
CompGeom for the set change which geoms belong to each set. This was
causing the Shown and Not_Shown sets to have components and results
removed from the Parasite Drag table. Additional considerations are
taken for Set type to ensure the table contains the same Geoms before
and after execution of the tool. This bug was identified by Brandon
Litherland.