Adds AVL Comparisons for Hershey Bar V&V Test Case
A function is added to read in an AVL strip force output. AVL
results for the AR10 Hershey Bar wing are read in and the
lift and drag distributions are plotted. AVL and XFoil results are
colored purple for consistency. The AVL input file and results
file are added to the scripts directory