Adds and Registers API Function to Update ParasiteDragMgr
A simple wrapper function is added to call ParasiteDragMgr.Update()
through the API. This function is then registered in ScriptMgr. This
function can be used to calculate and then get ParasiteDragMgr parms.
For example, SetParVal can be used to set velocity and altitude, from
which Reynold's Number per unit length can then be calculated by calling
the UpdateParasiteDrag function.