Convert 4-digit primary code path to use pseudo curve instead of pwcc.
The piecewise curve creator generated a nearly exact Bezier curve for a
4-digit airfoil.  However, since we had to create a PCHIP approximation
in order to convert to equal arc length parameterization, this exactness
was wasted.

The direct evaluation code path is much more simple and will be easier
to develop going forward.

The Bezier curve parameter caused LE clustering that lead to good
quality LE approximation in the PCHIP step.  This function is replaced
by calling the Cluster function..