Implements Available Control Surface Limitation for Groups
Instead of using the complete vector of all control surfaces, the
GetAvailableCSVec() is used to identify which control surfaces can be
added to a group. This is done in an effort to prevent a control surface
from being added to a group more than once. A control surface is still
able to belong to multiple groups.