Fixes Incorrect Implementation of White-Christoph Turbulent Cf Equation
Roughness was being used in the equation where gamma should have been.
Instead of adding a roughness input to the parasite drag table, the
gamma input is activated. In SetupHeatTransfer, the output string
was not being cleared, causing the Te/Tw input to be included in the
Taw/Tw input.