Sets Q Input to 1.2 for Wing Geoms with Jenkinson Tail FF Equation
This updates the GUI to only allow for a value of 1.2 for Q if
the Jenkinson Tail FF equation is indicated. This was already being done
withing the parasite drag calculation, but was not reflected in the GUI.
A TODO item is identified to change the Q input to an output for the
Jenkinson Tail FF equation.