BUGFIX: Adds Update Call to CreateSetupFile and Changes Call Order
First, a call to VSPAEROMgr.Update() is added to CreateSetupFile()
because the vectors of control surface and rotor groups were not being
setup correctly from the API. A TODO note is added to check if Update
calls can be removed from the VSPAERO analyses. The order of function
calls in VSPAEROMgr.Update() is adjusted to prevent the GUI browsers
from not showing the correct available and grouped control surfaces
until a second update when a vsp3 file is loaded.