Add design lift coefficient support to four digit airfoils (and modified)
This equation was derived in the computer algebra system maxima uses m,p in
terms of fractions of the chord -- not the x100 and x10 digits of the
NACA designation.

Theory of Wing Sections includes CLi with the definition of the camber
line and the attendant velocity increment.  However, those results are about
.02 off of those obtained with this exact result.

NACA Report 460 (the origin of the 4-digit system) includes CLi for the
camber lines in Table VIII on the last page of the report.  These values
were arrived at through analytical integration and match the results of
this equation essentially exactly.  The data points in the maxima program
are these points.

This commit also adds functions to perform the calculations as well as GUI
devices to support the choice of input mode (and the new input).