Add feather offset and axis control to propeller
Feather axis controls the airfoil x/c for the feather axis.  For
traditional constant speed props, this is 0.5 -- for traditional helicopter
rotors, this is 0.25.

Feather offset shifts the feather axis relative to the hub.  This will
make it easier to model some exotic prop designs.  This parameter will
usually not be used.