Fix crash when reading Lednicer airfoils.
This is largely a revert of 6e6fb63861a6e9e5e825a43454037e4376656af0 which
fixed a problem reading Selig airfoils that went slightly out of the
normal bounds of an airfoil (x[0,1], y[-1,1]).  Unfortunately, it also
rendered useless the determination of file type Selig vs. Lednicer.  That
check is based on the first numbers in the Lednicer file being the number
of points in the file -- a value expected to be well outside the bounds
of airfoil points.

This change attempts to satisfy both -- allow airfoil points wide berth
while still detecting Selig vs. Lednicer.