Remove un-needed ForceUpdate
This update was causing Unexpected ResetRemapID warnings because it
triggered a ResetRemapID before the closing ResetRemapID on read file.

It is believed that it is safe to defer the first update until after the
file is entirely read in -- so this fix should be correct.  However, if
something requires this update, there may be unintended consequences.